Failing step envs secret filtering test
