+++ _tests/integration/workflow_run_envs_test_secrets.yml
@@ -1,0 +1,2 @@
+envs:
+- SECRET_ERROR_MESSAGE: failure reason
